refactor(frontend): submission history query hookを分離
